int64, lint

+++ internal/app/orchestrator/orchestrator.go
@@ -370,7 +370,8 @@ func (o *orchestrator) onCacheEvicted(key string, resource cache.Resource) {
 	// TODO Potential for improvements here to handle the thundering herd
 	// problem: https://github.com/envoyproxy/xds-relay/issues/71
 	metrics.OrchestratorCacheEvictSubscope(o.scope, key).Counter(metrics.OrcheestratorCacheEvictCount).Inc(1)
-	metrics.OrchestratorCacheEvictSubscope(o.scope, key).Counter(metrics.OrchestratorOnCacheEvictedRequestCount).Inc(len(resource.Requests))
+	metrics.OrchestratorCacheEvictSubscope(o.scope, key).Counter(
+		metrics.OrchestratorOnCacheEvictedRequestCount).Inc(int64(len(resource.Requests)))
 	o.logger.With("aggregated_key", key).Debug(context.Background(), "cache eviction called")
 	o.downstreamResponseMap.deleteAll(resource.Requests)
 	o.upstreamResponseMap.delete(key)
